fix(tools): remove typo in tools name
